fix pr reviews
