+++ client/src/components/App/AppToolbar/AppToolbar.tsx
@@ -3,8 +3,9 @@ import { AppBar, Toolbar } from '@material-ui/core';
 import { Typography } from '@material-ui/core';
 import Tooltip from '@material-ui/core/Tooltip';
 
-import useAppToolbar from './useAppToolbar';
 import IsActiveToggle from 'commons/IsActiveToggle/IsActiveToggle';
+
+import useAppToolbar from './useAppToolbar';
 
 const toggleMessage = 'מה הסטטוס שלך?';
